+++ src/api/http.go
@@ -7,8 +7,6 @@ import (
 	"net"
 	"net/http"
 	"path/filepath"
-	"reflect"
-	"runtime"
 	"strings"
 	"time"
 	"unicode"
@@ -18,7 +16,6 @@ import (
 	"github.com/skycoin/skycoin/src/api/webrpc"
 	"github.com/skycoin/skycoin/src/daemon"
 	"github.com/skycoin/skycoin/src/readable"
-	"github.com/skycoin/skycoin/src/util/collections"
 	"github.com/skycoin/skycoin/src/util/file"
 	wh "github.com/skycoin/skycoin/src/util/http"
 	"github.com/skycoin/skycoin/src/util/logging"
@@ -37,16 +34,14 @@ const (
 	defaultWriteTimeout = time.Second * 60
 	defaultIdleTimeout  = time.Second * 120
 
-	// AllAPISets wildcard value to match all API methods
-	AllAPISets = "ALL"
-	// APIDefault endpoints available when nodes executed with no CLI args
-	APIDefault = "READ_ONLY"
-	// APIStatus endpoints offer (meta,runtime)data to dashboard and monitoring clients
-	APIStatus = "STATUS"
-	// APIWallet endpoints implement wallet interface
-	APIWallet = "WALLET"
-	// APISeed endpoints implement wallet interface
-	APISeed = "WALLET_SEED"
+	// EndpointsRead endpoints available when nodes executed with no CLI args
+	EndpointsRead = "READ"
+	// EndpointsStatus endpoints offer (meta,runtime)data to dashboard and monitoring clients
+	EndpointsStatus = "STATUS"
+	// EndpointsWallet endpoints implement wallet interface
+	EndpointsWallet = "WALLET"
+	// EndpointsWalletSeed endpoints implement wallet interface
+	EndpointsWalletSeed = "WALLET_SEED"
 )
 
 // Server exposes an HTTP API
@@ -68,7 +63,7 @@ type Config struct {
 	WriteTimeout         time.Duration
 	IdleTimeout          time.Duration
 	BuildInfo            readable.BuildInfo
-	EnabledAPISets       collections.StringSet
+	EnabledAPISets       map[string]struct{}
 }
 
 type muxConfig struct {
@@ -79,14 +74,7 @@ type muxConfig struct {
 	enableUnversionedAPI bool
 	disableCSP           bool
 	buildInfo            readable.BuildInfo
-	enabledAPISets       collections.StringSet
-}
-
-func (c muxConfig) isAPISetEnabled(APINames ...string) bool {
-	if len(APINames) == 0 {
-		return true
-	}
-	return c.enabledAPISets.ContainsAny(AllAPISets, APINames...)
+	enabledAPISets       map[string]struct{}
 }
 
 // HTTPResponse represents the http response struct
@@ -282,17 +270,25 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore, rpc *web
 		return handler
 	}
 
-	forAPISet := func(hf http.HandlerFunc, APINames ...string) http.HandlerFunc {
-		if len(APINames) == 0 {
-			return hf
+	forAPISet := func(f http.HandlerFunc, apiNames []string) http.HandlerFunc {
+		if len(apiNames) == 0 {
+			logger.Panic("apiNames should not be empty")
 		}
+
+		isEnabled := false
+
+		for _, k := range apiNames {
+			if _, ok := c.enabledAPISets[k]; ok {
+				isEnabled = true
+				break
+			}
+		}
+
 		return func(w http.ResponseWriter, r *http.Request) {
-			if !c.isAPISetEnabled(APINames...) {
-				funcName := runtime.FuncForPC(reflect.ValueOf(hf).Pointer()).Name()
-				logger.Infof("Handler %s not executed because API set %v not enabled", funcName, APINames)
-				wh.Error403(w, "")
+			if isEnabled {
+				f(w, r)
 			} else {
-				hf(w, r)
+				wh.Error403(w, "")
 			}
 		}
 	}
@@ -355,10 +351,10 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore, rpc *web
 	webHandlerV1("/version", versionHandler(c.buildInfo))
 
 	// get set of unspent outputs
-	webHandlerV1("/outputs", forAPISet(getOutputsHandler(gateway), APIDefault))
+	webHandlerV1("/outputs", forAPISet(getOutputsHandler(gateway), []string{EndpointsRead}))
 
 	// get balance of addresses
-	webHandlerV1("/balance", forAPISet(getBalanceHandler(gateway), APIDefault))
+	webHandlerV1("/balance", forAPISet(getBalanceHandler(gateway), []string{EndpointsRead}))
 
 	// Wallet interface
 
@@ -366,7 +362,7 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore, rpc *web
 	// Method: GET
 	// Args:
 	//      id - Wallet ID [required]
-	webHandlerV1("/wallet", forAPISet(walletGet(gateway), APIWallet))
+	webHandlerV1("/wallet", forAPISet(walletGet(gateway), []string{EndpointsWallet}))
 
 	// Loads wallet from seed, will scan ahead N address and
 	// load addresses till the last one that have coins.
@@ -375,16 +371,16 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore, rpc *web
 	//     seed: wallet seed [required]
 	//     label: wallet label [required]
 	//     scan: the number of addresses to scan ahead for balances [optional, must be > 0]
-	webHandlerV1("/wallet/create", forAPISet(walletCreate(gateway), APIWallet))
+	webHandlerV1("/wallet/create", forAPISet(walletCreate(gateway), []string{EndpointsWallet}))
 
-	webHandlerV1("/wallet/newAddress", forAPISet(walletNewAddresses(gateway), APIWallet))
+	webHandlerV1("/wallet/newAddress", forAPISet(walletNewAddresses(gateway), []string{EndpointsWallet}))
 
 	// Returns the confirmed and predicted balance for a specific wallet.
 	// The predicted balance is the confirmed balance minus any pending
 	// spent amount.
 	// GET arguments:
 	//      id: Wallet ID
-	webHandlerV1("/wallet/balance", forAPISet(walletBalanceHandler(gateway), APIWallet))
+	webHandlerV1("/wallet/balance", forAPISet(walletBalanceHandler(gateway), []string{EndpointsWallet}))
 
 	// Sends coins&hours to another address.
 	// POST arguments:
@@ -393,121 +389,121 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore, rpc *web
 	//  dst: Destination address
 	//  Returns total amount spent if successful, otherwise error describing
 	//  failure status.
-	webHandlerV1("/wallet/spend", forAPISet(walletSpendHandler(gateway), APIWallet))
+	webHandlerV1("/wallet/spend", forAPISet(walletSpendHandler(gateway), []string{EndpointsWallet}))
 
 	// Creates a transaction from a wallet
-	webHandlerV1("/wallet/transaction", forAPISet(createTransactionHandler(gateway), APIWallet))
+	webHandlerV1("/wallet/transaction", forAPISet(createTransactionHandler(gateway), []string{EndpointsWallet}))
 
 	// GET Arguments:
 	//      id: Wallet ID
 	// Returns all pending transanction for all addresses by selected Wallet
-	webHandlerV1("/wallet/transactions", forAPISet(walletTransactionsHandler(gateway), APIWallet))
+	webHandlerV1("/wallet/transactions", forAPISet(walletTransactionsHandler(gateway), []string{EndpointsWallet}))
 
 	// Update wallet label
 	// POST Arguments:
 	//     id: wallet id
 	//     label: wallet label
-	webHandlerV1("/wallet/update", forAPISet(walletUpdateHandler(gateway), APIWallet))
+	webHandlerV1("/wallet/update", forAPISet(walletUpdateHandler(gateway), []string{EndpointsWallet}))
 
 	// Returns all loaded wallets
 	// returns sensitive information
-	webHandlerV1("/wallets", forAPISet(walletsHandler(gateway), APIWallet))
+	webHandlerV1("/wallets", forAPISet(walletsHandler(gateway), []string{EndpointsWallet}))
 
 	// Returns wallets directory path
-	webHandlerV1("/wallets/folderName", forAPISet(getWalletFolder(gateway), APIWallet))
+	webHandlerV1("/wallets/folderName", forAPISet(getWalletFolder(gateway), []string{EndpointsWallet}))
 
 	// Generate wallet seed
 	// GET Arguments:
 	//     entropy: entropy bitsize.
-	webHandlerV1("/wallet/newSeed", forAPISet(newWalletSeed(gateway), APIWallet))
+	webHandlerV1("/wallet/newSeed", forAPISet(newWalletSeed(gateway), []string{EndpointsWallet}))
 
 	// Gets seed of wallet of given id
 	// GET Arguments:
 	//     id: wallet id
 	//     password: wallet password
-	webHandlerV1("/wallet/seed", forAPISet(walletSeedHandler(gateway), APISeed))
+	webHandlerV1("/wallet/seed", forAPISet(walletSeedHandler(gateway), []string{EndpointsWalletSeed}))
 
 	// unload wallet
 	// POST Argument:
 	//         id: wallet id
-	webHandlerV1("/wallet/unload", forAPISet(walletUnloadHandler(gateway), APIWallet))
+	webHandlerV1("/wallet/unload", forAPISet(walletUnloadHandler(gateway), []string{EndpointsWallet}))
 
 	// Encrypts wallet
 	// POST arguments:
 	//     id: wallet id
 	//     password: wallet password
 	// Returns an encrypted wallet json without sensitive data
-	webHandlerV1("/wallet/encrypt", forAPISet(walletEncryptHandler(gateway), APIWallet))
+	webHandlerV1("/wallet/encrypt", forAPISet(walletEncryptHandler(gateway), []string{EndpointsWallet}))
 
 	// Decrypts wallet
 	// POST arguments:
 	//     id: wallet id
 	//     password: wallet password
-	webHandlerV1("/wallet/decrypt", forAPISet(walletDecryptHandler(gateway), APIWallet))
+	webHandlerV1("/wallet/decrypt", forAPISet(walletDecryptHandler(gateway), []string{EndpointsWallet}))
 
 	// Blockchain interface
 
-	webHandlerV1("/blockchain/metadata", forAPISet(blockchainMetadataHandler(gateway), APIStatus, APIDefault))
-	webHandlerV1("/blockchain/progress", forAPISet(blockchainProgressHandler(gateway), APIStatus, APIDefault))
+	webHandlerV1("/blockchain/metadata", forAPISet(blockchainMetadataHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
+	webHandlerV1("/blockchain/progress", forAPISet(blockchainProgressHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
 
 	// get block by hash or seq
-	webHandlerV1("/block", forAPISet(blockHandler(gateway), APIDefault))
+	webHandlerV1("/block", forAPISet(blockHandler(gateway), []string{EndpointsRead}))
 	// get blocks in specific range
-	webHandlerV1("/blocks", forAPISet(blocksHandler(gateway), APIDefault))
+	webHandlerV1("/blocks", forAPISet(blocksHandler(gateway), []string{EndpointsRead}))
 	// get last N blocks
-	webHandlerV1("/last_blocks", forAPISet(lastBlocksHandler(gateway), APIDefault))
+	webHandlerV1("/last_blocks", forAPISet(lastBlocksHandler(gateway), []string{EndpointsRead}))
 
 	// Network stats interface
-	webHandlerV1("/network/connection", forAPISet(connectionHandler(gateway), APIStatus, APIDefault))
-	webHandlerV1("/network/connections", forAPISet(connectionsHandler(gateway), APIStatus, APIDefault))
-	webHandlerV1("/network/defaultConnections", forAPISet(defaultConnectionsHandler(gateway), APIStatus, APIDefault))
-	webHandlerV1("/network/connections/trust", forAPISet(trustConnectionsHandler(gateway), APIStatus, APIDefault))
-	webHandlerV1("/network/connections/exchange", forAPISet(exchgConnectionsHandler(gateway), APIStatus, APIDefault))
+	webHandlerV1("/network/connection", forAPISet(connectionHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
+	webHandlerV1("/network/connections", forAPISet(connectionsHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
+	webHandlerV1("/network/defaultConnections", forAPISet(defaultConnectionsHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
+	webHandlerV1("/network/connections/trust", forAPISet(trustConnectionsHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
+	webHandlerV1("/network/connections/exchange", forAPISet(exchgConnectionsHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
 
 	// Transaction handler
 
 	// get set of pending transactions
-	webHandlerV1("/pendingTxs", forAPISet(pendingTxnsHandler(gateway), APIDefault))
+	webHandlerV1("/pendingTxs", forAPISet(pendingTxnsHandler(gateway), []string{EndpointsRead}))
 	// get txn by txid
-	webHandlerV1("/transaction", forAPISet(transactionHandler(gateway), APIDefault))
+	webHandlerV1("/transaction", forAPISet(transactionHandler(gateway), []string{EndpointsRead}))
 
 	// parse and verify transaction
-	webHandlerV2("/transaction/verify", verifyTxnHandler(gateway))
+	webHandlerV2("/transaction/verify", forAPISet(verifyTxnHandler(gateway), []string{EndpointsRead}))
 
 	// Health check handler
-	webHandlerV1("/health", forAPISet(healthHandler(c, csrfStore, gateway), APIStatus, APIDefault))
+	webHandlerV1("/health", forAPISet(healthHandler(c, csrfStore, gateway), []string{EndpointsRead, EndpointsStatus}))
 
 	// Returns transactions that match the filters.
 	// Method: GET
 	// Args:
 	//     addrs: Comma separated addresses [optional, returns all transactions if no address is provided]
 	//     confirmed: Whether the transactions should be confirmed [optional, must be 0 or 1; if not provided, returns all]
-	webHandlerV1("/transactions", forAPISet(getTransactions(gateway), APIDefault))
+	webHandlerV1("/transactions", forAPISet(getTransactions(gateway), []string{EndpointsRead}))
 	// inject a transaction into network
-	webHandlerV1("/injectTransaction", forAPISet(injectTransaction(gateway), APIDefault))
-	webHandlerV1("/resendUnconfirmedTxns", forAPISet(resendUnconfirmedTxns(gateway), APIDefault))
+	webHandlerV1("/injectTransaction", forAPISet(injectTransaction(gateway), []string{EndpointsRead}))
+	webHandlerV1("/resendUnconfirmedTxns", forAPISet(resendUnconfirmedTxns(gateway), []string{EndpointsRead}))
 	// get raw tx by txid.
-	webHandlerV1("/rawtx", forAPISet(getRawTxn(gateway), APIDefault))
+	webHandlerV1("/rawtx", forAPISet(getRawTxn(gateway), []string{EndpointsRead}))
 
 	// UxOut api handler
 
 	// get uxout by id.
-	webHandlerV1("/uxout", forAPISet(getUxOutByID(gateway), APIDefault))
+	webHandlerV1("/uxout", forAPISet(getUxOutByID(gateway), []string{EndpointsRead}))
 	// get all the address affected uxouts.
-	webHandlerV1("/address_uxouts", forAPISet(getAddrUxOuts(gateway), APIDefault))
+	webHandlerV1("/address_uxouts", forAPISet(getAddrUxOuts(gateway), []string{EndpointsRead}))
 
-	webHandlerV2("/address/verify", http.HandlerFunc(forAPISet(addressVerify, APIDefault)))
+	webHandlerV2("/address/verify", forAPISet(addressVerify, []string{EndpointsRead}))
 
 	// Explorer handler
 
 	// get set of pending transactions
-	webHandlerV1("/explorer/address", forAPISet(getTransactionsForAddress(gateway), APIDefault))
+	webHandlerV1("/explorer/address", forAPISet(getTransactionsForAddress(gateway), []string{EndpointsRead}))
 
-	webHandlerV1("/coinSupply", forAPISet(coinSupply(gateway), APIStatus, APIDefault))
+	webHandlerV1("/coinSupply", forAPISet(coinSupply(gateway), []string{EndpointsRead}))
 
-	webHandlerV1("/richlist", forAPISet(getRichlist(gateway), APIStatus, APIDefault))
+	webHandlerV1("/richlist", forAPISet(getRichlist(gateway), []string{EndpointsRead}))
 
-	webHandlerV1("/addresscount", forAPISet(getAddressCount(gateway), APIStatus, APIDefault))
+	webHandlerV1("/addresscount", forAPISet(getAddressCount(gateway), []string{EndpointsRead}))
 
 	return mux
 }