+++ src/api/http_test.go
@@ -9,20 +9,23 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"github.com/skycoin/skycoin/src/util/collections"
 )
 
 const configuredHost = "127.0.0.1:6420"
 
-func defaultMuxConfig(enabledAPINames ...string) muxConfig {
-	if len(enabledAPINames) == 0 {
-		enabledAPINames = append(enabledAPINames, APIDefault)
-	}
+var allAPISetsEnabled = map[string]struct{}{
+	EndpointsRead:       struct{}{},
+	EndpointsStatus:     struct{}{},
+	EndpointsWallet:     struct{}{},
+	EndpointsWalletSeed: struct{}{},
+}
+
+func defaultMuxConfig() muxConfig {
 	return muxConfig{
 		host:           configuredHost,
 		appLoc:         ".",
 		disableCSP:     true,
-		enabledAPISets: collections.NewStringSet(enabledAPINames...),
+		enabledAPISets: allAPISetsEnabled,
 	}
 }