+++ src/skycoin/skycoin.go
@@ -321,11 +321,11 @@ func (c *Coin) ConfigureDaemon() daemon.Config {
 	dc.Visor.GenesisCoinVolume = c.config.Node.GenesisCoinVolume
 	dc.Visor.DBPath = c.config.Node.DBPath
 	dc.Visor.Arbitrating = c.config.Node.Arbitrating
-	dc.Visor.EnableWalletAPI = c.config.Node.EnableWalletAPI
 	dc.Visor.WalletDirectory = c.config.Node.WalletDirectory
-	dc.Visor.EnableSeedAPI = c.config.Node.EnableSeedAPI
+	_, dc.Visor.EnableWalletAPI = c.config.Node.enabledAPISets[api.EndpointsWallet]
+	_, dc.Visor.EnableSeedAPI = c.config.Node.enabledAPISets[api.EndpointsWalletSeed]
 
-	dc.Gateway.EnabledAPISets = c.config.Node.WebInterfaceAPISets
+	_, dc.Gateway.EnableWalletAPI = c.config.Node.enabledAPISets[api.EndpointsWallet]
 
 	// Initialize wallet default crypto type
 	cryptoType, err := wallet.CryptoTypeFromString(c.config.Node.WalletCryptoType)
@@ -352,7 +352,7 @@ func (c *Coin) createGUI(d *daemon.Daemon, host string) (*api.Server, error) {
 		ReadTimeout:          c.config.Node.ReadTimeout,
 		WriteTimeout:         c.config.Node.WriteTimeout,
 		IdleTimeout:          c.config.Node.IdleTimeout,
-		EnabledAPISets:       c.config.Node.WebInterfaceAPISets,
+		EnabledAPISets:       c.config.Node.EnabledAPISets,
 		BuildInfo: readable.BuildInfo{
 			Version: c.config.Build.Version,
 			Commit:  c.config.Build.Commit,